Allow passing None to split_args_with
Passing None to split() causes splitting to occur on _any_ form of
whitespace, which is different from splitting on ' '. Previous to this
commit, passing None was equivalent to saying, "don't split args".

With this commit, the default for this is now '' (Which isn't a valid
value to split()), freeing None to be passed on to split() instead
without breaking backwards-compatibility.