Dont test the type of NoneType elements in recurse_check_structure()
recurse_check_structure() is used to check plugin configurations.
However, checking the type of NoneType elements makes no sense because
then a plugin could only ever be configured with the key set to None
(because anything else would be of a different type) which defeats the
purpose of having that configurable parameter.

Of course it'll be up to plugins to check the validity of these keys.