Use rocket for the webserver
Bottle's support for various different servers is nice, and giving the
end user a choice of webserver is advantageous, but unfortunately stands
in the way of generalized SSL support.

Every server implements SSL differently, using SSL will require custom
code to start the server. Implementing SSL for all the webserver
supported by Bottle would be unfeasable, thus we are left with making a
choice of webserver anyway. Keeping the choice of server open for those
cases where SSL isn't used then becomes a rather marginal benefit with
potential confusing results for end-users.

I've settled on Rocket for now because Rocket:
* Is well documented
* Works with Python 2.x and 3.x without a need for 2to3
* Supports SSL out of the box without needing to hook in custom classes
  or generate SSL contexts by hand
* Makes it dead simple to run multiple sockets, so we can have a plain
  HTTP as well as an HTTPS socket simultaniously, rather than making it
  an HTTP- or HTTPS-only choice for the end user