Add bot command to generate SSL certificate directly from Err
This adds the command '!generate certificate' which will generate a
self-signed certificate directly from Err, instead of relying on a
seperate script to do it. Relies on PyOpenSSL but checks to make
sure it's installed and simply gives a warning when it's not, so
PyOpenSSL doesn't become a hard dependency.