Fix replaces the call to XEP_0045.getRoomForm with XEP_0045.getRoomConfig
The latter uses the api present in xep_0004 plugin. The method
getRoomForm used the old version of the plugin. Apart from that
the functionality of this two methods was for all practical reasons
identical.

It seems that the plugins are capable of loading their own
dependencies.
In case of xep_0045 it is xep_0004. When I removed the line from err,
nothing happened. However, leaving the line in place for the sake of
explicitness. (And because at the moment I don't know if some part
of the bot uses it directly.)

Fixes #233 and #234.