Only configure XMPP MUC when having owner affiliation
When joining an XMPP MUC, we should only attempt to configure it when
having the 'owner' affiliation, as only owners may configure a room.

Room configuration should happen mainly when the bot is creating a new
room (by joining a room that did not exist yet), otherwise other
people may not join the room yet. (See section 10.1 of XEP-0045).

With this change, room configuration is still attempted at all times,
but only when our affiliation is actually of type owner.

Thanks to @apophys for doing the groundwork on this.

Closes #237: RFE: Configure room only when necessary