Move decorators into a dedicated submodule
I've mainly done this because Sphinx doesn't pick up on functions inside
`errbot/__init__.py` so @botcmd and @re_botcmd were going undocumented
in the docs. This change ensures they can be referred to easily in the
docs.

While at it, the docstrings for these two decorators, as well as
@webhook, have been greatly expanded.