Remove HipChat html support.
HipChat does not support incoming html-im [0] and has only
rudimentary support for HTML if you send it through the HTTP
API. There are however two main drawbacks with this (as documented on
the API page [1] for that call):

* Special HipChat features such as @mentions, emoticons, and image
  previews are NOT supported when using this format.
* There may be a slight delay before messages appear in the room.

Furthermore, the API is subject to pretty strict rate limiting. Personal
access tokens only get 100 requests per 5 minutes [2]. It would be very
trivial to exhaust the available number of API calls, blocking the API
from functioning and preventing further messages from going out.
It would also prevent the MUC management from working, since it's
subject to the same API limits.

With these limitations currently in place, it is my belief that simply
sending the xhtml-im part like with any other XMPP server, letting it
get ignored by HipChat until such a time that they do support it (they
mention it is a wish they have, see the official reaction in the
comments on [0]) is better than making it sort-of-work by intercepting
messages and routing them over to the HTTP API.

[0]: http://help.hipchat.com/forums/138883-suggestions-ideas/suggestions/2974079-support-xhtml-im-send-text-only-notifications-to
[1]: https://www.hipchat.com/docs/apiv2/method/send_room_notification
[2]: https://www.hipchat.com/docs/apiv2/rate_limiting