[slack] Be more helpful about PY3-related SlackClient import errors
Python 3 support on SlackClient is very recent (there's no released
version on PyPI with support for it at the time of this writing) and the
package lacks any automated testing. I believe the chance is high that
we will see some Python 3 related issues with this library the future.

While catching a SyntaxError at import time only covers a small amount
of possible cases, it at least covers the `print` instead of `print()`
case that is happening with the currently released version on PyPI.
Instead of exploding with only a stack trace, this will at least give
users a better understanding of what is going on (and make them more
likely to blame the right people :)).