Implement generic reconnection logic
Some back-ends currently implement reconnection logic because the
underlying libraries (irc, SleekXMPP) support this. This commit
introduces a generic, back-end agnostic reconnection system with
exponential backoff and randomized jitter that any back-end can use,
irrespective of whether or not the underlying library offers this.

The reason for this change is to support reconnection on Slack and any
other future back-ends that don't support this natively.

It has been a conscious decision to implement this logic in Err
rather than in python-slackclient directly because doing so would make
it impossible to trigger `disconnect_callback` on disconnections. This
is because the IRC and SleekXMPP libraries are threaded and implement a
callback for this event, but SlackClient is not, so if it handled
reconnections internally there would be no way to signal Err to trigger
`disconnect_callback`.