Prevent old plugin to break on __init__
Old plugins could override __init__ with no parameters as yapsy was only
able to do that.

Now that we pass a bot ref for callbacks, if the plugin incorrectly
forwards the parameters, it breaks.

This commit prevents this by checking the __init__ function and adapt,
emitting a warning if it had to work around it.