Ported backends as plugins.
This allows third party backends to be run by err.

backends have the same .plug as a normal err plugin.
The backend has to be a subclass of ErrBot.

The port had a ripple effect across the codebase that had to be cleaned
nup, especially tests were we were doing dodgy things like globals
(this is because yapsy loads plugins in a virtual modules).

err can now be started with a backend name with -b or --backend.

Another new command line parameter -l to list the available backends
will be added in another PR.