Friendly error when joining rooms on Slack and user_is_bot
This will ensure users see a helpful error message in the logs when
entering rooms in CHATROOM_PRESENCE on Slack and using a bot account, as
well returning similar messages when issuing !room commands to the bot.

It was a conscious decision to handle these events in the chatroom
plugin instead of ignoring the error entirely so that plugin authors can
still check for this occurence. A plugin might rely on being able to
create or join a room, so silently swallowing the error message is not a
good approach.