Avoid restart loop on Telegram
On Telegram, the server keeps messages for the bot stored until you ask
for messages with an offset higher than that of what it has stored
(Telegram's way of confirming message receipt).

This means that in the naive event where we ask for messages without an
offset (as happened on bot startup), we get old messages if we didn't
confirm them on the previous getUpdates() run. In the case of issuing a
bot command such as !restart, the bot shuts down before asking for new
updates with a higher offset, resulting in it receiving the
unconfirmed !restart command again.

It immediately restarts, and gets it again. And so on in and endless
loop.

By storing the offset in persistent storage and loading it at bot
startup, we can avoid this situation.