Changes order of development deps in setup.py.
This is due to a bug/feature in setuptools:

https://bitbucket.org/pypa/setuptools/issues/196/tests_require-pytest-pytest-cov-breaks

Essentially, `pytest-xdist` is being matched as the `xdist` version of pytest
instead of recognizing it as a separate package; as a result, `pytest-xdist` is
fulfilling the role of `pyest`, when in reality it does not actually provide
that functionality.

The current workaround is to change the order of operations so that the
`pytest` package is the last to be installed after any other `pytest-*`
packages have been processed.