Ensure arg_botcmd error and help output goes into chat
By default, argparse will print errors, usage and help text directly to
stdout/stderr. It will also raise SystemExit to make the program stop
with an appropriate error code.

While this is appropriate for a CLI program, it is horrible for Err. The
SystemExit causes some strange behaviour where it doesn't make the bot
exit, but does prevent some processing (!restart no longer works, for
example) and the printing to stdout/stderr means you get no feedback
when issuing commands with incorrect arguments.

With these changes, all this output is nicely returned to the user,
instead.

(Bonus: The command name is used in the usage as well, instead of the
default `sys.arv[0]` which resulted in one of `err.py`, `errbot` or
`run_tests.py`.)