Implements a filter for core plugins.
Some bundled plugins are not useful for everybody and creates noise in
the Help system and potential security issues if they are not configured
correctly.

This adds a CORE_PLUGINS entry to the config.py so users can select the
set of core plugins they need at config time.

If it is None or unspecified, all the coreplugins are started.