Fix IRC channel join hang on startup
The IRC backend is currently hanging on startup when it tries to join
the rooms, because the join room call blocks other processing of server
events somehow.

It eventually times out raising an exception about not being able to
join the channel, then continues to run the rest of bot startup.

When run in it's own thread, the bot successfully connects without
issue and properly joins channels.