[Slack] Don't try to treat URIs as person mentions
Slack modifies URIs by putting them in between angle brackets, which is
how the bot received them. The callback_mention code was picking up on
these and trying to treat them as identifiers, which would break.

We now still try to build an identifier from them, but if
build_identifier thinks it's not a valid one (it has all the logic for
this!) we just ignore it and assume it wasn't an identifier after all.

Fixes #622.