Merge ACCESS_CONTROLS_DEFAULT into ACCESS_CONTROLS rules
This ensures that all the settings defined in ACCESS_CONTROLS_DEFAULT
are applied to each command listed in ACCESS_CONTROLS. If a certain
command should to override the default set in ACCESS_CONTROLS_DEFAULT,
it must be explicitly set up.

This is more natural and more intuitive, removing the need to specify
the default for every single command where you define other rules.