Close storage of BotRepoManager on shutdown
This was a tricky one. I could reproduce this on one machine but not
another. A curious difference between the two was that they were using
different dbm implementations.

On the problematic one:
    data/repomgr.db.db: Berkeley DB (Hash, version9, native byte-order)

On the one that wouldn't reproduce the issue:
   data/repomgr.db: GNU dbm 1.x or ndbm database, little endian, old

Note also the Berkeley DB implementation adds an extra `.db` to the
filename which the GNU dbm implementation does not.

It appears the Berkeley version needs a clean `.close()` or it won't
persist updated values, GNU dbm updates them immediately.

Fixes #650