Notify admins about plugin startup errors during bot startup
Previously, these messages would get logged but not otherwise be
visible because there was no connection to the chat network yet.

This means that if a plugin failed to start, you would have to look
into the logs to see what happened.

Now we store any errors which occurred during bot startup and send
them to the admins once we have successfully connected.