Allow for disabling slack markdown conversion
When a plugin is being used and that plugin is
only being made to work with slack there is no
need to go through the markdown conversion since
the bot can natively just write the supported
slack format in the first place.

So this allows for that conversion to be turned
off (it is left on by default) via a new RAW_BODY
configuration option.