Don't import from test backend, use own TextPerson
This avoids attempting to import pytest which is not a required dependency
for errbot.

For now this is effectively a duplication of the original class in the
test backend to avoid entangling these together. In the future we may wish
to refactor this and provide a common, shared base between these two
backends but I think that should be part of a larger effort where we may
wish to emulate all aspects of a chat network..