Statified slack_client
We had a problem with the Slack Backend because a lot of indentifier related
manupulations needs a live slack client but pickling and unpickling
need static functions.

In this PR I push slack_client as a global for the module and cascaded
the change through all the identifier creations and transformations.

This simplifies the code and made the pickling/unpickling code less
finicky at the expense of the structure of the code (basically the whole
backend is a huge static blob now).