Revert "Fix the docs requirements file"
This reverts commit d081bf0937094a20f65964a2a528d080605154cf.

ReadTheDocs attempts to build the sphinx documentation with a working
directory set to the root of the errbot project/repo itself. That means
it needs `.` to refer to errbot's own `setup.py` file, not `./..`.

It's a bit confusing because it means you can't `pip install -r
requirements.txt` from the docs folder, but have to do so from one level
up. :( I don't see a good way around that though (if anyone has a better
solution, I'm all ears).