fix(repo_manager): Use utc timestamps (#1)
In a fresh instance of errbot with no previous storage data, the `LAST_UPDATE` time is set to 0 to indicate it's never been checked. However, when you search or install a new repo, it checks to see if should do an update, but since it was using `datetime.fromtimestamp` it would fail (at least on Windows, this might be OS specific) since 0 is an invalid value. We can preserve the intended behavior by instead switching to `datetime.utcfromtimestamp` and `datetime.utcnow` functions.