Use ShlexArgParser in lieu of SeparatorArgParser
Some of the room commands in chatRoom.py use SeparatorArgParser which doesn’t parse room names correctly that contain spaces and are wrapped in quotes. For example the command:

> room create "this is a new room"
< Created the room "this

> room create another new room
< Created the room another

By using ShlexArgParser the rooms with spaces and wrapped in quotes will be parsed correctly. An arg without quotes or spaces will continue to be parsed correctly. For example:

> room create "this is a new room"
< Created the room this is a new room

> room list
< I'm currently in these rooms:
    this is a new room (group)
    <output cut for brevity>

> room create this_is_a_room_with_no_spaces_or_quotes@blah.com
< Created the room this_is_a_room_with_no_spaces_or_quotes@blah.com

> room list
< I'm currently in these rooms:
    this_is_a_room_with_no_spaces_or_quotes@blah.com (group)
    this is a new room (group)
    <output cut for brevity>