Only use one underscore
Two underscores looked great and more like the python way to do things
until I saw that some fonts were putting the two underscores into a
single underscore, and that would be confusing to end-users if the docs
said the same thing.  By using a single underscore and using the `'str'
in str` method both double-underscore and single-underscore values will
be acceptable.