docs: Fix typographical issues in setup guide (#1475)
* docs: Make list out of packaged versions

It looks like the plain text is written to be a list, but doesn't follow
the reStructuredText format for an unordered list.

* docs: Use :code: role for command execution

Primarily fix an issue where Docutils Smart Quotes transformation will
turn double-dash `--` into en-dash `–`. This might be confusing to new
users who copy and paste from the text but don't realize the CLI option
should be double-dash.