Don't use string representation for acl matches.
While fixing the documentation, I noticed that rooms are using str(room)
for acl matches. `__str__` is meant to be a human-readable object
representation which could be something other than the room ID -
although as no one spotted this I guess that in most cases it is.

This commit makes it so that rooms have an aclattr, just like people do.
This can then be used to store a suitable version of the room for ACL
matches.

I've tried to maintain backwards compatability in two ways:

  1. aclattr defaults to the old behaviour str(self)
  2. the getter can detect if the room (Occupant) is a room or a str,
     and if it's a str use that instead. I know for the Matrix backend
     I'd incorrectly been returning a string and it still worked - if
     other backends are doing that it'll throw an exception without this
     mitigation.